Wire ClearSettings into main Settings menu; fix animationSpeed Float crash
Re-exported debloat/hide-ai-features patch now wires the Cleargram
Settings entry into the main Settings menu (PeerInfoScreen/BUILD,
PeerInfoScreen.swift, PeerInfoScreenSettingsActions.swift,
PeerInfoSettingsItems.swift) instead of Debug-only access.

Fixed EXC_BREAKPOINT crash when toggling any setting: Postbox's
AdaptedPostboxEncoder has no Float overload in its keyed container
(only Double/Int32/Int64/Bool/String/Data), so Float fell through to
singleValueContainer() -> preconditionFailure(). Changed animationSpeed
to Double in ClearConfigSettings, ClearConfig, ClearSettingsController.

+++ src/swift/ClearGram/DebugSettingsUI/ClearSettingsController.swift
@@ -19,7 +19,7 @@ private enum ClearSettingsEntry: ItemListNodeEntry {
     case hideStories(PresentationTheme, Bool)
     case hideAiFeatures(PresentationTheme, Bool)
     case doubleTapDelay(PresentationTheme, Int32)
-    case animationSpeed(PresentationTheme, Float)
+    case animationSpeed(PresentationTheme, Double)
 
     var section: ItemListSectionId {
         switch self {
@@ -140,7 +140,7 @@ public func clearSettingsController(context: AccountContext) -> ViewController {
             tf.text = String(format: "%.1f", ClearConfig.animationSpeed)
         }
         alert.addAction(UIAlertAction(title: "Set", style: .default) { _ in
-            if let text = alert.textFields?.first?.text, let value = Float(text) {
+            if let text = alert.textFields?.first?.text, let value = Double(text) {
                 let _ = ClearConfig.update(accountManager: context.sharedContext.accountManager) { var s = $0; s.animationSpeed = value; return s }.start()
             }
         })
